requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
pydantic==2.10.6
pydantic_core==2.27.2
PyPDF2==3.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
wheel==0.45.1
